# PERF: use GEOSGeom_getExtent for GEOS>=3.11 in bounds ufunc