--- a/CHANGES.txt
+++ b/CHANGES.txt
@@ -8,6 +8,8 @@ Relevant changes compared to 2.0a1:
 
 - Renamed the ``tolerance`` keyword to ``max_segment_length`` in the
   ``segmentize`` function.
+- Added use of ``GEOSGeom_getExtent`` to speed up bounds calculations for
+  GEOS >= 3.11.
 
 2.0a1 (2022-08-03)
 ------------------