xref https://github.com/libgeos/geos/pull/554 / https://github.com/libgeos/geos/pull/555

This gives a modest speedup for simple geometries:

```
In [1]: import shapely

In [2]: polygons = shapely.polygons(np.random.randn(1_000_000, 3, 2))

In [3]: %timeit shapely.total_bounds(polygons)
64.9 ms ± 3.13 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   # <- main
46.6 ms ± 2.31 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   # <- PR
```