@jorisvandenbossche 2022-09-10
> Please add a note to the changelog, but otherwise this looks good, thanks!

Given that this is a new function in 2.0.0, is that worth including? (all top-level vectorized functions have a better performance compared to doing that before manually)