@brendan-ward 2022-09-10
It is sometimes helpful to know when we've wrapped a particular GEOS function when tracing issues.

What I meant was something like this under 2.0b1:

```
- added use of GEOSGeom_getExtent to speed up bounds calculations for GEOS >= 3.11
```

But it is minor, so if you don't think it fits the intent of changes listed for 2.0 / 2.0b1, it is fine to leave it out.